+++ src/language/interpreter.rs
@@ -1020,316 +1020,315 @@ impl Sqrt for i64 {
     }
 }
 
+extern crate test;
+
 #[cfg(test)]
 mod tests {
+
     use super::*;
     use approx::AbsDiffEq;
     use ndarray::array;
     use std::str::FromStr;
+    use test::Bencher;
 
-    #[test]
-    fn compute_elementwise_add_0() {
-        let mut env = Environment::new();
-        env.insert(
-            "t",
-            array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
+    /// Creates a benchmark test for the interpreter
+    /// The test does the following:
+    ///  1. Parses $glenside_str as glenside expression
+    ///  2. Creates a new Environment from the vector of (key, value) pairs
+    ///  3. Calls the interpreter with the glenside expression and env,
+    ///         and passes the value to check_correct
+    /// $test_name: the name of the created benchmark test
+    /// $glenside_str: A string containing the Glenside program
+    /// $env: A vector of 2-tuples of key, value pairs to put into the environment
+    /// $check_correct: A closure with arguments (value) that checks for correctness
+    macro_rules! benchmark_test {
+        ($test_name:ident, $glenside_str: expr, $env: expr, $check_correct: expr) => {
+            #[bench]
+            fn $test_name(b: &mut Bencher) {
+                let mut env = Environment::new();
+                for (key, value) in $env.into_iter() {
+                    env.insert(key, value);
+                }
 
-        let expr = RecExpr::<Language>::from_str(
-            "(compute elementwise-add
-              (access (access-tensor t) 0)
-             )",
-        )
-        .unwrap();
+                let expr = RecExpr::<Language>::from_str($glenside_str).unwrap();
 
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 0);
-                assert_eq!(
-                    tensor,
-                    array![[1 + -5 + -9, -2 + 6 + 10], [3 + 0 + 11, 0 + 8 + 12]].into_dyn()
-                );
+                b.iter(|| {
+                    // use black box to prevent compiler optimizations
+                    let expr = test::black_box(&expr);
+                    let env = test::black_box(&env);
+
+                    let value = interpret(&expr, expr.as_ref().len() - 1, &env);
+                    $check_correct(value);
+                });
             }
-            _ => panic!(),
-        }
+        };
     }
 
-    #[test]
-    fn compute_elementwise_mul_0() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_elementwise_add_0,
+        "(compute elementwise-add
+        (access (access-tensor t) 0)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute elementwise-mul
-              (access (access-tensor t) 0)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 0);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[1 * -5 * -9, -2 * 6 * 10], [3 * 0 * 11, 0 * 8 * 12]].into_dyn()
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 0);
+                    assert_eq!(
+                        tensor,
+                        array![[1 + -5 + -9, -2 + 6 + 10], [3 + 0 + 11, 0 + 8 + 12]].into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_reduce_sum_0() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_elementwise_mul_0,
+        "(compute elementwise-mul
+        (access (access-tensor t) 0)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute reduce-sum
-              (access (access-tensor t) 0)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 0);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    ndarray::arr0(1 + -2 + 3 + 0 + -5 + 6 + 0 + 8 + -9 + 10 + 11 + 12).into_dyn()
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 0);
+                    assert_eq!(
+                        tensor,
+                        array![[1 * -5 * -9, -2 * 6 * 10], [3 * 0 * 11, 0 * 8 * 12]].into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_reduce_sum_1() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_reduce_sum_0,
+        "(compute reduce-sum
+        (access (access-tensor t) 0)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute reduce-sum
-              (access (access-tensor t) 1)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 1);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![1 + -2 + 3 + 0, -5 + 6 + 0 + 8, -9 + 10 + 11 + 12].into_dyn()
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 0);
+                    assert_eq!(
+                        tensor,
+                        ndarray::arr0(1 + -2 + 3 + 0 + -5 + 6 + 0 + 8 + -9 + 10 + 11 + 12)
+                            .into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_reduce_sum_2() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_reduce_sum_1,
+        "(compute reduce-sum
+        (access (access-tensor t) 1)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute reduce-sum
-              (access (access-tensor t) 2)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 2);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[1 + -2, 3 + 0], [-5 + 6, 0 + 8], [-9 + 10, 11 + 12]].into_dyn()
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 1);
+                    assert_eq!(
+                        tensor,
+                        array![1 + -2 + 3 + 0, -5 + 6 + 0 + 8, -9 + 10 + 11 + 12].into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_reduce_sum_3() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_reduce_sum_2,
+        "(compute reduce-sum
+        (access (access-tensor t) 2)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute reduce-sum
-              (access (access-tensor t) 3)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 3);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 2);
+                    assert_eq!(
+                        tensor,
+                        array![[1 + -2, 3 + 0], [-5 + 6, 0 + 8], [-9 + 10, 11 + 12]].into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_relu_0() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_reduce_sum_3,
+        "(compute reduce-sum
+        (access (access-tensor t) 3)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute relu
-              (access (access-tensor t) 0)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 0);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[[1, 0], [3, 0]], [[0, 6], [0, 8]], [[0, 10], [11, 12]],].into_dyn(),
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 3);
+                    assert_eq!(
+                        tensor,
+                        array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],]
+                            .into_dyn(),
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_relu_1() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_relu_0,
+        "(compute relu
+        (access (access-tensor t) 0)
+        )",
+        vec![(
             "t",
             array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute relu
-              (access (access-tensor t) 2)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(access_axis, 2);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[[1, 0], [3, 0]], [[0, 6], [0, 8]], [[0, 10], [11, 12]],].into_dyn(),
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 0);
+                    assert_eq!(
+                        tensor,
+                        array![[[1, 0], [3, 0]], [[0, 6], [0, 8]], [[0, 10], [11, 12]],].into_dyn(),
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_dot_product_0() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_relu_1,
+        "(compute relu
+        (access (access-tensor t) 2)
+        )",
+        vec![(
+            "t",
+            array![[[1, -2], [3, 0]], [[-5, 6], [0, 8]], [[-9, 10], [11, 12]],].into_dyn(),
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(access_axis, 2);
+                    assert_eq!(
+                        tensor,
+                        array![[[1, 0], [3, 0]], [[0, 6], [0, 8]], [[0, 10], [11, 12]],].into_dyn(),
+                    );
+                }
+                _ => panic!(),
+            }
+        }
+    );
+
+    benchmark_test!(
+        compute_dot_product_0,
+        "(compute dot-product
+        (access (access-tensor t) 0)
+        )",
+        vec![(
             "t",
             // 3 x 2 x 2
             array![[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute dot-product
-              (access (access-tensor t) 0)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor.shape(), &[] as &[usize]);
-                assert_eq!(access_axis, 0);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    ndarray::arr0(1 * 5 * 9 + 2 * 6 * 10 + 3 * 7 * 11 + 4 * 8 * 12).into_dyn()
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor.shape(), &[] as &[usize]);
+                    assert_eq!(access_axis, 0);
+                    assert_eq!(
+                        tensor,
+                        ndarray::arr0(1 * 5 * 9 + 2 * 6 * 10 + 3 * 7 * 11 + 4 * 8 * 12).into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn compute_dot_product_1() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        compute_dot_product_1,
+        "(compute dot-product
+        (access (access-tensor t) 1)
+        )",
+        vec![(
             "t",
             // 3 x 2 x 2
             array![[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]],].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(compute dot-product
-              (access (access-tensor t) 1)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor.shape(), &[3]);
-                assert_eq!(access_axis, 1);
-                assert_eq!(
+        )],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![11, 5 * 7 + 8 * 6, 9 * 11 + 10 * 12].into_dyn()
-                );
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor.shape(), &[3]);
+                    assert_eq!(access_axis, 1);
+                    assert_eq!(
+                        tensor,
+                        array![11, 5 * 7 + 8 * 6, 9 * 11 + 10 * 12].into_dyn()
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
     #[test]
     fn compute_dot_product_2() {
@@ -1363,47 +1362,46 @@ mod tests {
         }
     }
 
-    #[test]
-    fn access_cartesian_product() {
-        let mut env = Environment::new();
-        env.insert(
-            "t0",
-            // 3 x 2 x 2
-            array![[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]],].into_dyn(),
-        );
-        env.insert(
-            "t1",
-            // 2 x 2 x 2
-            array![[[13, 14], [15, 16]], [[17, 18], [19, 20]]].into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "(access-cartesian-product
-              (access (access-tensor t0) 2)
-              (access (access-tensor t1) 2)
-             )",
-        )
-        .unwrap();
-
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor.shape(), &[3, 2, 2, 2, 2, 2]);
-                assert_eq!(access_axis, 4);
-                assert_eq!(
-                    tensor.slice(s![0, 0, 0, 0, .., ..]),
-                    array![[1, 2], [13, 14]]
-                );
-                assert_eq!(
-                    tensor.slice(s![2, 0, 1, 0, .., ..]),
-                    array![[9, 10], [17, 18]]
-                );
+    benchmark_test!(
+        access_cartesian_product,
+        "(access-cartesian-product
+        (access (access-tensor t0) 2)
+        (access (access-tensor t1) 2)
+        )",
+        vec![
+            (
+                "t0",
+                // 3 x 2 x 2
+                array![[[1, 2], [3, 4]], [[5, 6], [7, 8]], [[9, 10], [11, 12]],].into_dyn(),
+            ),
+            (
+                "t1",
+                // 2 x 2 x 2
+                array![[[13, 14], [15, 16]], [[17, 18], [19, 20]]].into_dyn(),
+            )
+        ],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor.shape(), &[3, 2, 2, 2, 2, 2]);
+                    assert_eq!(access_axis, 4);
+                    assert_eq!(
+                        tensor.slice(s![0, 0, 0, 0, .., ..]),
+                        array![[1, 2], [13, 14]]
+                    );
+                    assert_eq!(
+                        tensor.slice(s![2, 0, 1, 0, .., ..]),
+                        array![[9, 10], [17, 18]]
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
+
     #[test]
     fn access() {
         let mut env = Environment::new();
@@ -1432,52 +1430,48 @@ mod tests {
         interpret(&expr, expr.as_ref().len() - 1, &env);
     }
 
-    #[test]
-    fn access_windows() {
-        let mut env = Environment::new();
-        env.insert(
+    benchmark_test!(
+        access_windows,
+        "(access-windows
+            (access (access-tensor t) 3)
+            (shape 3 2 2)
+            (shape 1 1 1)
+        )",
+        vec![(
             "t",
             array![
                 [[1., 2., 3.], [4., 5., 6.], [7., 8., 9.]],
                 [[10., 11., 12.], [13., 14., 15.], [16., 17., 18.]],
                 [[19., 20., 21.], [22., 23., 24.], [25., 26., 27.]],
             ]
-            .into_dyn(),
-        );
-
-        let expr = RecExpr::<Language>::from_str(
-            "
-             (access-windows
-              (access (access-tensor t) 3)
-              (shape 3 2 2)
-              (shape 1 1 1)
-             )",
-        )
-        .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(a) => {
-                assert_eq!(a.access_axis, 3);
-                assert_eq!(a.tensor.shape(), &[1, 2, 2, 3, 2, 2]);
-                assert_eq!(
-                    a.tensor.slice(s![0, 0, 0, .., .., ..]),
-                    array![
-                        [[1., 2.], [4., 5.]],
-                        [[10., 11.], [13., 14.]],
-                        [[19., 20.], [22., 23.]],
-                    ]
-                );
-                assert_eq!(
-                    a.tensor.slice(s![0, 1, 0, .., .., ..]),
-                    array![
-                        [[4., 5.], [7., 8.]],
-                        [[13., 14.], [16., 17.]],
-                        [[22., 23.], [25., 26.]],
-                    ]
-                );
+            .into_dyn()
+        )],
+        |value| {
+            match value {
+                Value::Access(a) => {
+                    assert_eq!(a.access_axis, 3);
+                    assert_eq!(a.tensor.shape(), &[1, 2, 2, 3, 2, 2]);
+                    assert_eq!(
+                        a.tensor.slice(s![0, 0, 0, .., .., ..]),
+                        array![
+                            [[1., 2.], [4., 5.]],
+                            [[10., 11.], [13., 14.]],
+                            [[19., 20.], [22., 23.]],
+                        ]
+                    );
+                    assert_eq!(
+                        a.tensor.slice(s![0, 1, 0, .., .., ..]),
+                        array![
+                            [[4., 5.], [7., 8.]],
+                            [[13., 14.], [16., 17.]],
+                            [[22., 23.], [25., 26.]],
+                        ]
+                    );
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
     #[test]
     fn shape() {
@@ -1692,38 +1686,36 @@ mod tests {
         };
     }
 
-    #[test]
-    fn access_pad() {
-        let mut env = Environment::new();
-        env.insert("t", array![[1., 2.], [3., 4.]].into_dyn());
-
-        let expr =
-            RecExpr::<Language>::from_str("(access-pad (access-tensor t) zero-padding 0 2 4)")
-                .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(
+    benchmark_test!(
+        access_pad,
+        "(access-pad (access-tensor t) zero-padding 0 2 4)",
+        vec![("t", array![[1., 2.], [3., 4.]].into_dyn())],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![
-                        [0., 0.],
-                        [0., 0.],
-                        [1., 2.],
-                        [3., 4.],
-                        [0., 0.],
-                        [0., 0.],
-                        [0., 0.],
-                        [0., 0.]
-                    ]
-                    .into_dyn()
-                );
-                assert_eq!(access_axis, 0);
+                    access_axis,
+                }) => {
+                    assert_eq!(
+                        tensor,
+                        array![
+                            [0., 0.],
+                            [0., 0.],
+                            [1., 2.],
+                            [3., 4.],
+                            [0., 0.],
+                            [0., 0.],
+                            [0., 0.],
+                            [0., 0.]
+                        ]
+                        .into_dyn()
+                    );
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
     #[test]
     fn compute_reduce_max_0() {
@@ -1928,83 +1920,80 @@ mod tests {
         }
     }
 
-    #[test]
-    fn access_pair_0() {
-        let mut env = Environment::new();
-        env.insert("a", array![[1, 2], [3, 4]].into_dyn());
-        env.insert("b", array![[5, 6], [7, 8]].into_dyn());
-
-        let expr = RecExpr::<Language>::from_str(
-            "(access-pair (access (access-tensor a) 0) (access (access-tensor b) 0))",
-        )
-        .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor.shape(), [2, 2, 2]);
-                assert_eq!(
+    benchmark_test!(
+        access_pair_0,
+        "(access-pair (access (access-tensor a) 0) (access (access-tensor b) 0))",
+        vec![
+            ("a", array![[1, 2], [3, 4]].into_dyn()),
+            ("b", array![[5, 6], [7, 8]].into_dyn())
+        ],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[[1, 2], [3, 4]], [[5, 6], [7, 8]]].into_dyn()
-                );
-                assert_eq!(access_axis, 0);
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor.shape(), [2, 2, 2]);
+                    assert_eq!(
+                        tensor,
+                        array![[[1, 2], [3, 4]], [[5, 6], [7, 8]]].into_dyn()
+                    );
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn access_pair_1() {
-        let mut env = Environment::new();
-        env.insert("a", array![[1, 2], [3, 4]].into_dyn());
-        env.insert("b", array![[5, 6], [7, 8]].into_dyn());
-
-        let expr = RecExpr::<Language>::from_str(
-            "(access-pair (access (access-tensor a) 1) (access (access-tensor b) 1))",
-        )
-        .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor.shape(), [2, 2, 2]);
-                assert_eq!(
+    benchmark_test!(
+        access_pair_1,
+        "(access-pair (access (access-tensor a) 1) (access (access-tensor b) 1))",
+        vec![
+            ("a", array![[1, 2], [3, 4]].into_dyn()),
+            ("b", array![[5, 6], [7, 8]].into_dyn())
+        ],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[[1, 2], [5, 6]], [[3, 4], [7, 8]]].into_dyn()
-                );
-                assert_eq!(access_axis, 1);
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor.shape(), [2, 2, 2]);
+                    assert_eq!(
+                        tensor,
+                        array![[[1, 2], [5, 6]], [[3, 4], [7, 8]]].into_dyn()
+                    );
+                    assert_eq!(access_axis, 1);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn access_pair_2() {
-        let mut env = Environment::new();
-        env.insert("a", array![[1, 2], [3, 4]].into_dyn());
-        env.insert("b", array![[5, 6], [7, 8]].into_dyn());
-
-        let expr = RecExpr::<Language>::from_str(
-            "(access-pair (access (access-tensor a) 2) (access (access-tensor b) 2))",
-        )
-        .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor.shape(), [2, 2, 2]);
-                assert_eq!(
+    benchmark_test!(
+        access_pair_2,
+        "(access-pair (access (access-tensor a) 2) (access (access-tensor b) 2))",
+        vec![
+            ("a", array![[1, 2], [3, 4]].into_dyn()),
+            ("b", array![[5, 6], [7, 8]].into_dyn())
+        ],
+        |value| {
+            match value {
+                Value::Access(Access {
                     tensor,
-                    array![[[1, 5], [2, 6]], [[3, 7], [4, 8]]].into_dyn()
-                );
-                assert_eq!(access_axis, 2);
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor.shape(), [2, 2, 2]);
+                    assert_eq!(
+                        tensor,
+                        array![[[1, 5], [2, 6]], [[3, 7], [4, 8]]].into_dyn()
+                    );
+                    assert_eq!(access_axis, 2);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
     #[test]
     #[should_panic]
@@ -2663,49 +2652,47 @@ mod tests {
         }
     }
 
-    #[test]
-    fn access_concatenate_0() {
-        let mut env = Environment::new();
-        env.insert("t", array![[1, 2]].into_dyn());
-        env.insert("n", array![[1, 2], [3, 4]].into_dyn());
-
-        let expr = RecExpr::<Language>::from_str(
-            "(access-concatenate (access (access-tensor t) 0) (access (access-tensor n) 0) 0)",
-        )
-        .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor, array![[1, 2], [1, 2], [3, 4]].into_dyn());
-                assert_eq!(access_axis, 0);
+    benchmark_test!(
+        access_concatenate_0,
+        "(access-concatenate (access (access-tensor t) 0) (access (access-tensor n) 0) 0)",
+        vec![
+            ("t", array![[1, 2]].into_dyn()),
+            ("n", array![[1, 2], [3, 4]].into_dyn())
+        ],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor, array![[1, 2], [1, 2], [3, 4]].into_dyn());
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn access_concatenate_1() {
-        let mut env = Environment::new();
-        env.insert("t", array![[1], [2]].into_dyn());
-        env.insert("n", array![[1, 2], [3, 4]].into_dyn());
-
-        let expr = RecExpr::<Language>::from_str(
-            "(access-concatenate (access (access-tensor t) 0) (access (access-tensor n) 0) 1)",
-        )
-        .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor, array![[1, 1, 2], [2, 3, 4]].into_dyn());
-                assert_eq!(access_axis, 0);
+    benchmark_test!(
+        access_concatenate_1,
+        "(access-concatenate (access (access-tensor t) 0) (access (access-tensor n) 0) 1)",
+        vec![
+            ("t", array![[1], [2]].into_dyn()),
+            ("n", array![[1, 2], [3, 4]].into_dyn())
+        ],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor, array![[1, 1, 2], [2, 3, 4]].into_dyn());
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
     #[test]
     #[should_panic]
@@ -2735,65 +2722,59 @@ mod tests {
         interpret(&expr, expr.as_ref().len() - 1, &env);
     }
 
-    #[test]
-    fn access_slice_0() {
-        let mut env = Environment::new();
-        env.insert("t", array![[1, 2], [3, 4]].into_dyn());
-
-        let expr =
-            RecExpr::<Language>::from_str("(access-slice (access (access-tensor t) 0) 0 0 1)")
-                .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor, array![[1, 2]].into_dyn());
-                assert_eq!(access_axis, 0);
+    benchmark_test!(
+        access_slice_0,
+        "(access-slice (access (access-tensor t) 0) 0 0 1)",
+        vec![("t", array![[1, 2], [3, 4]].into_dyn())],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor, array![[1, 2]].into_dyn());
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn access_slice_1() {
-        let mut env = Environment::new();
-        env.insert("t", array![[1, 2], [3, 4]].into_dyn());
-
-        let expr =
-            RecExpr::<Language>::from_str("(access-slice (access (access-tensor t) 0) 0 0 2)")
-                .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor, array![[1, 2], [3, 4]].into_dyn());
-                assert_eq!(access_axis, 0);
+    benchmark_test!(
+        access_slice_1,
+        "(access-slice (access (access-tensor t) 0) 0 0 2)",
+        vec![("t", array![[1, 2], [3, 4]].into_dyn())],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor, array![[1, 2], [3, 4]].into_dyn());
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
-    #[test]
-    fn access_slice_2() {
-        let mut env = Environment::new();
-        env.insert("t", array![[1, 2], [3, 4]].into_dyn());
-
-        let expr =
-            RecExpr::<Language>::from_str("(access-slice (access (access-tensor t) 0) 1 1 2)")
-                .unwrap();
-        match interpret(&expr, expr.as_ref().len() - 1, &env) {
-            Value::Access(Access {
-                tensor,
-                access_axis,
-            }) => {
-                assert_eq!(tensor, array![[2], [4]].into_dyn());
-                assert_eq!(access_axis, 0);
+    benchmark_test!(
+        access_slice_2,
+        "(access-slice (access (access-tensor t) 0) 1 1 2)",
+        vec![("t", array![[1, 2], [3, 4]].into_dyn())],
+        |value| {
+            match value {
+                Value::Access(Access {
+                    tensor,
+                    access_axis,
+                }) => {
+                    assert_eq!(tensor, array![[2], [4]].into_dyn());
+                    assert_eq!(access_axis, 0);
+                }
+                _ => panic!(),
             }
-            _ => panic!(),
         }
-    }
+    );
 
     #[test]
     #[should_panic]
